refactor(vm): split vm_call_method_compiled.rs into cohesive submodules (§7-8)
Pure relocation: 2668-line file split into 6 flat sibling files with no
logic changes. Indivisible giants (_interpret.rs: 669L, _mut.rs: 500L)
each go alone in their own file per convention.

- vm_call_method_compiled.rs (~302L): protect-block ops + wrap-chain check
- vm_call_method_compiled_cache.rs (~349L): multi-arg type keys, method
  resolve cache, populate-uncompiled, dispatch-compiled, fast-cache populate
- vm_call_method_compiled_coerce.rs (~316L): quanthash/map-hash/IO
  coercions, iterator construction
- vm_call_method_compiled_interpret.rs (~669L): try_compiled_method_or_interpret
  wrapper + inner giant (indivisible)
- vm_call_method_compiled_io.rs (~547L): IO::Handle state, text/byte output,
  read methods
- vm_call_method_compiled_mut.rs (~500L): try_compiled_method_mut_or_interpret
  (near-indivisible giant)
