+++ src/vm.rs
@@ -136,6 +136,11 @@ mod vm_call_helpers;
 mod vm_call_light;
 mod vm_call_light_typed;
 mod vm_call_method_compiled;
+mod vm_call_method_compiled_cache;
+mod vm_call_method_compiled_coerce;
+mod vm_call_method_compiled_interpret;
+mod vm_call_method_compiled_io;
+mod vm_call_method_compiled_mut;
 mod vm_call_method_mut_ops;
 mod vm_call_method_ops;
 mod vm_call_named;
