+++ src/vm/vm_call_method_compiled.rs
@@ -88,2372 +88,6 @@ impl Interpreter {
         Ok(Some(result))
     }
 
-    /// Try compiled method fast path; fall back to interpreter.
-    ///
-    /// Wrapper that preserves the caller's env `self`: the inner dispatch can run
-    /// an interpreted instance method, which binds `self` in env without the
-    /// save/restore that `call_method_with_values` performs. Without this, a bare
-    /// stringification (`"$obj"` → `StringConcat` → this) leaves the caller's
-    /// `self` pointing at `$obj`, breaking a later `self` read in an enclosing
-    /// nested sub (which resolves `self` from env via `GetSelfOrNoSelf`).
-    /// Build a per-positional-arg type key for the sound multi-resolution cache.
-    /// Returns `None` (do not cache) when any arg is value-/identity-dependent or
-    /// autothreads (`Junction`), or is a container/named-pair arg, so dispatch can
-    /// not be keyed on the positional types alone.
-    fn multi_arg_type_keys(args: &[Value]) -> Option<Vec<crate::symbol::Symbol>> {
-        let mut keys = Vec::with_capacity(args.len());
-        for a in args {
-            let key = match a {
-                Value::Instance { class_name, .. } => *class_name,
-                Value::Junction { .. }
-                | Value::Mixin(..)
-                | Value::Scalar(_)
-                | Value::ContainerRef(_)
-                | Value::Pair(..)
-                | Value::ValuePair(..)
-                | Value::Capture { .. } => return None,
-                other => {
-                    crate::symbol::Symbol::intern(crate::runtime::utils::value_type_name(other))
-                }
-            };
-            keys.push(key);
-        }
-        Some(keys)
-    }
-
-    /// Whether a `(class, method)` is a MULTI whose dispatch is purely type+arity
-    /// based — i.e. the resolved candidate is a function of the receiver class +
-    /// method + positional arg types, so it is safe to cache in
-    /// `multi_resolve_cache`. False for non-multi methods (the existing
-    /// `method_resolve_cache` handles those) and for any multi with a value-/
-    /// identity-dependent candidate (`where` / literal / subset / `:D`/`:U` smiley /
-    /// coercion). Memoized per `(class, method)`.
-    fn multi_dispatch_type_cacheable(
-        &mut self,
-        class_sym: crate::symbol::Symbol,
-        method_sym: crate::symbol::Symbol,
-        class_name: &str,
-        method_name: &str,
-    ) -> bool {
-        if let Some(&c) = self.multi_type_cacheable.get(&(class_sym, method_sym)) {
-            return c;
-        }
-        let mro = self.class_mro(class_name);
-        let mut any_multi = false;
-        let mut value_dependent = false;
-        'outer: for cn in &mro {
-            let Some(overloads) = self.registry().get_method_overloads(cn, method_name) else {
-                continue;
-            };
-            for def in &overloads {
-                if def.is_multi {
-                    any_multi = true;
-                }
-                for pd in &def.param_defs {
-                    if pd.where_constraint.is_some() || pd.literal_value.is_some() {
-                        value_dependent = true;
-                        break 'outer;
-                    }
-                    if let Some(tc) = &pd.type_constraint {
-                        // `:D`/`:U`/`:_` smiley or `Int(Str)` coercion => value/identity
-                        // dependent; a subset type carries an implicit `where`.
-                        if tc.contains(':') || tc.contains('(') {
-                            value_dependent = true;
-                            break 'outer;
-                        }
-                        let base = tc.split(['[', ' ']).next().unwrap_or(tc.as_str());
-                        if self.registry().subsets.contains_key(base) {
-                            value_dependent = true;
-                            break 'outer;
-                        }
-                    }
-                }
-            }
-        }
-        let cacheable = any_multi && !value_dependent;
-        self.multi_type_cacheable
-            .insert((class_sym, method_sym), cacheable);
-        cacheable
-    }
-
-    /// Resolve a method, consulting the sound multi-resolution cache for a
-    /// type+arity-deterministic multi (avoids the per-call MRO/specificity walk).
-    /// Non-multi / uncacheable / un-keyable calls resolve fresh (the non-multi
-    /// caches live at the call sites). An AMBIGUOUS multi resolution is never
-    /// cached — it must re-raise `X::Multi::Ambiguous` on every call (a cache hit
-    /// would not set `dispatch_ambiguous`).
-    fn resolve_method_cached(
-        &mut self,
-        cn: &str,
-        method: &str,
-        class_sym: crate::symbol::Symbol,
-        method_sym: crate::symbol::Symbol,
-        args: &[Value],
-        target: &Value,
-    ) -> Option<(String, std::sync::Arc<crate::runtime::MethodDef>)> {
-        if let Some(arg_keys) = Self::multi_arg_type_keys(args)
-            && self.multi_dispatch_type_cacheable(class_sym, method_sym, cn, method)
-        {
-            let mkey = (class_sym, method_sym, arg_keys);
-            if let Some(hit) = self.multi_resolve_cache.get(&mkey) {
-                return hit.clone();
-            }
-            let resolved = loan_env!(
-                self,
-                resolve_method_with_owner_invocant(cn, method, args, target)
-            );
-            let resolved_arc = resolved.map(|(o, d)| (o, std::sync::Arc::new(d)));
-            if !self.dispatch_ambiguous {
-                self.multi_resolve_cache.insert(mkey, resolved_arc.clone());
-            }
-            return resolved_arc;
-        }
-        loan_env!(
-            self,
-            resolve_method_with_owner_invocant(cn, method, args, target)
-        )
-        .map(|(o, d)| (o, std::sync::Arc::new(d)))
-    }
-
-    pub(super) fn try_compiled_method_or_interpret(
-        &mut self,
-        target: Value,
-        method: &str,
-        args: Vec<Value>,
-    ) -> Result<Value, RuntimeError> {
-        let saved_self = self.get_env_with_main_alias("self");
-        let result = self.try_compiled_method_or_interpret_inner(target, method, args);
-        match saved_self {
-            Some(s) => self.set_env_with_main_alias("self", s),
-            None => {
-                self.env_mut().remove("self");
-            }
-        }
-        result
-    }
-
-    fn try_compiled_method_or_interpret_inner(
-        &mut self,
-        target: Value,
-        method: &str,
-        args: Vec<Value>,
-    ) -> Result<Value, RuntimeError> {
-        // Native default construction: `Foo.new(...)` for a simple user-defined
-        // class is pure data assembly (named args + attribute defaults), so the
-        // Interpreter builds the instance directly instead of routing through the
-        // interpreter's generic constructor dispatch (lever A: shrink method-call
-        // interpreter fallback).
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && let Some(result) = loan_env!(self, try_native_default_construct(*class_name, &args))
-        {
-            // An exception type with a user-defined `message` method needs that
-            // method run once at construction and cached into the `message`
-            // attribute, exactly as the interpreter's `dispatch_new` caller does
-            // (`materialize_exception_message_in_result`). For built-in exceptions
-            // and non-exceptions this is a no-op (no user `message` method).
-            let cn = class_name.resolve();
-            let result = self.materialize_exception_message_in_result(result);
-            // Native construction of an attribute-only class is pure data assembly
-            // (named args + defaults; writes nothing to the caller env, Slice 6.3).
-            // BUT if the class has a `submethod BUILD`/`TWEAK`, the native path runs
-            // that phase, whose body can mutate a captured-outer caller lexical
-            // (`my $n; submethod TWEAK { $n++ }`) — so that construction is NOT
-            // env-pure: leave the dispatch impure so the call site reconciles the
-            // caller's slot (Slice F, `reconcile_locals_from_env_at_site`). A
-            // user `message` method materialized above can likewise run user code,
-            // so treat that as impure too.
-            let runs_message =
-                (cn == "Exception" || cn.starts_with("X::") || cn.starts_with("CX::"))
-                    && self.has_user_method(&cn, "message");
-            self.method_dispatch_pure = !self.mro_has_build_or_tweak(&cn) && !runs_message;
-            return result;
-        }
-        // Native built-in construction: `Buf`/`Blob` (byte overlay), `utf8`/
-        // `utf16` (code units), `Uni` (codepoints), `Version`/`Duration`/
-        // `StrDistance`/`Stash`/empty-instance handles — pure data builds the Interpreter
-        // performs directly instead of routing through `dispatch_new`.
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && let Some(result) =
-                crate::runtime::Interpreter::try_native_builtin_construct(*class_name, &args)
-        {
-            self.method_dispatch_pure = true;
-            return result;
-        }
-        // Native QuantHash construction: `Set`/`SetHash`/`Bag`/`BagHash`/`Mix`/
-        // `MixHash`.new(...) — pure element counting + optional parameterized
-        // type-check / container-metadata tag, no env / registry / user code.
-        // Built directly via the single `try_native_quanthash_construct` impl the
-        // interpreter's `dispatch_new` also delegates to.
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && let Some(result) =
-                self.try_native_quanthash_construct_for_package(*class_name, &args)
-        {
-            self.method_dispatch_pure = true;
-            return result;
-        }
-        // Native aggregate construction: `Array`/`List`/`Positional`/`array`/
-        // `Hash`/`Map`.new(...) — shaped-dim parsing + parameterized type check +
-        // container-metadata tag, no env / registry / user code. Built directly via
-        // the single `try_native_array_construct` / `try_native_hash_construct`
-        // impls the interpreter's `dispatch_new` also delegates to.
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && let Some(result) =
-                self.try_native_aggregate_construct_for_package(*class_name, &args)
-        {
-            self.method_dispatch_pure = true;
-            return result;
-        }
-        // Native IO::Path family construction: `IO::Path`/`IO::Path::Unix`/`::Win32`/
-        // `::Cygwin`/`::QNX`.new(...) — pure path-string assembly (positional path
-        // or basename/dirname/volume/CWD/SPEC pairs), no FS / cwd / env / user code
-        // (registry reads + a one-time SPEC-subclass registration, which the VM
-        // owns). Built via the single `build_io_path_instance` impl the
-        // interpreter's `dispatch_new` arm also delegates to.
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && let Some(result) = self.try_native_io_path_construct(*class_name, &args)
-        {
-            self.method_dispatch_pure = true;
-            return result;
-        }
-        // Native Failure construction: `Failure.new($exception?)` — pure data
-        // assembly reading only VM-owned state (`$!` from env, the exception's MRO
-        // from the registry), no FS / process / user code. Built via the single
-        // `build_native_failure_value` impl the interpreter's
-        // `dispatch_new_and_constructors` arm also delegates to.
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && class_name.resolve() == "Failure"
-        {
-            self.method_dispatch_pure = true;
-            return Ok(self.build_native_failure_value(&args));
-        }
-        // Native Seq construction: `Seq.new($iterator?)` — registers the
-        // iterator into the VM-owned predictive/deferred carrier tables (no
-        // eager pull, no FS / process / user code). Built via the single
-        // `try_native_seq_construct` impl the interpreter's `dispatch_new` arm
-        // also delegates to. (A user subclass `class S is Seq` resolves to its
-        // own class name and is left to the interpreter.)
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && class_name.resolve() == "Seq"
-        {
-            self.method_dispatch_pure = true;
-            return Ok(self.try_native_seq_construct(&args));
-        }
-        // Native IO::Socket::INET construction: `IO::Socket::INET.new(...)` —
-        // the real bind/connect writes only VM-owned `io_handles` state (same
-        // shape as the native `IO::Path.open`). Built via the single
-        // `dispatch_socket_inet_new` impl the interpreter's `dispatch_new` arm
-        // also delegates to. (A user subclass resolves to its own class name and
-        // is left to the interpreter.)
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && class_name.resolve() == "IO::Socket::INET"
-        {
-            self.method_dispatch_pure = true;
-            return self.dispatch_socket_inet_new(&args);
-        }
-        // Native built-in *class* method (a pure type-object method other than
-        // `.new`, e.g. `Instant.from-posix`) — built directly instead of routing
-        // through the interpreter's class-method dispatch.
-        if let Value::Package(class_name) = &target
-            && let Some(result) = crate::runtime::Interpreter::try_native_builtin_class_method(
-                *class_name,
-                method,
-                &args,
-            )
-        {
-            self.method_dispatch_pure = true;
-            return result;
-        }
-        if let Value::Instance { class_name, .. } = &target {
-            let class = class_name.resolve();
-            // Interpreter-native pure-handle IO dispatch (PLAN.md ③ native IO PR-C/PR-D):
-            // resolve `IO::Handle`'s state-only methods (close/tell/eof/seek/
-            // opened/t and the Tier-1 setters/getters chomp/nl-out/out-buffer/
-            // encoding/native-descriptor) in the Interpreter through its own `io_handles`
-            // handle, *before* the generic native-method fallback below would
-            // otherwise pre-empt them. Returns `None` (falling through to that
-            // fallback unchanged) for any receiver/method/args it cannot handle
-            // natively, so this is behavior-invariant.
-            if let Some(result) = self.try_native_io_handle_method(&target, method, &args) {
-                return result;
-            }
-            // Interpreter-native text output to a File+UTF8 `IO::Handle` (print/put/say/
-            // print-nl): the write touches only handle state (PR-D Tier-2a).
-            // Stdout/Stderr (need emit_output) and non-UTF8 File fall through.
-            if let Some(result) = self.try_native_io_handle_output(&target, method, &args) {
-                return result;
-            }
-            // Interpreter-native raw byte output to a File `IO::Handle` (write/spurt):
-            // raw file write, no buffering/encoding (PR-D Tier-2c). Stdout/Stderr
-            // and a non-UTF8 spurt of a Str fall through.
-            if let Some(result) = self.try_native_io_handle_byte_output(&target, method, &args) {
-                return result;
-            }
-            // Interpreter-native line read from a File+UTF8 `IO::Handle` (get): reads via
-            // the handle's record reader (PR-D read side). ArgFiles/Stdin/non-UTF8
-            // (which need @*ARGS / decode) fall through.
-            if let Some(result) = self.try_native_io_handle_read(&target, method, &args) {
-                return result;
-            }
-            // Interpreter-native pure-lexical `IO::Path` methods (`.parent`/`.add`/
-            // `.basename`/`.sibling`/`.parts`/`.extension`/…): derive a new
-            // path/string/bool purely from the receiver's attributes, with no
-            // filesystem / cwd / env dependency (ledger §D). Single impl shared with
-            // the interpreter's `native_io_path`. Filesystem / cwd-relative forms and
-            // `child :secure` return `None` and fall through to the native fork below.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) =
-                    Self::try_io_path_lexical(&class, &attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native `.absolute` / `.relative` (path + cwd, lexical — no
-            // filesystem; the VM owns env/cwd). Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) =
-                    self.try_io_path_cwd_method(&attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native filesystem `stat`-only file tests / accessors
-            // (`e`/`f`/`d`/…/`s`/`modified`): resolve the path against the VM-owned
-            // cwd, then `stat` only — no `io_handles`, no content read (ledger §D).
-            // Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) = self.try_io_path_fs_stat(&attributes.as_map(), method)
-            {
-                return result;
-            }
-            // Interpreter-native whole-file content reads (`slurp`/`lines`/`words`):
-            // read the file + split/decode; no `io_handles` (ledger §D). Single impl
-            // shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) =
-                    self.try_io_path_content_read(&attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native single-path filesystem mutations (`spurt`/`mkdir`/
-            // `rmdir`/`unlink`/`chmod`): one-shot syscall, no `io_handles` (ledger
-            // §D). Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) =
-                    self.try_io_path_fs_mutate(&attributes.as_map(), &class, method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native `open`: allocate an `io_handles` entry and return
-            // the `IO::Handle`. The VM owns `io_handles`, so this is a native
-            // dispatch (ledger §D ③). Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) = self.try_io_path_open(&attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native two-path FS ops (`copy`/`rename`/`move`/`symlink`/
-            // `link`): resolve both paths against the VM-owned cwd, one-shot syscall,
-            // no `io_handles` (ledger §D). Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) =
-                    self.try_io_path_two_path_op(&attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native `comb`: read the file then comb the content (no
-            // `io_handles`; ledger §D). Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) = self.try_io_path_comb(&attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // A user-defined subclass of a builtin type may override an inherited
-            // native method (e.g. `class IO::Blob is IO::Handle { method get {…} }`).
-            // The user override must win, so do not take the native fork when the
-            // class (via its MRO) provides its own method of this name.
-            if self.is_native_method(&class, method) && !self.has_user_method(&class, method) {
-                // TODO: compile to bytecode — Instance native-method fork (ledger §1).
-                crate::vm::vm_stats::record_method_fallback(method);
-                return loan_env!(self, call_method_with_values(target, method, args));
-            }
-        }
-        // CARRIER: MOP pseudo-methods (reflection; no bytecode form). See ledger §C.
-        // Pseudo-methods must always go through the interpreter which handles
-        // them specially — never intercept via the compiled fast path.
-        if matches!(
-            method,
-            "DEFINITE" | "WHAT" | "WHO" | "HOW" | "WHY" | "WHICH" | "WHERE" | "VAR"
-        ) {
-            crate::vm::vm_stats::record_method_fallback(method);
-            return loan_env!(self, call_method_with_values(target, method, args));
-        }
-        // Private method fast path: resolve private candidate and run compiled code
-        // when caller context clearly allows direct dispatch.
-        if method.starts_with('!') {
-            let class_name = match &target {
-                Value::Instance { class_name, .. } => Some(class_name.resolve()),
-                Value::Package(name) => Some(name.resolve()),
-                _ => None,
-            };
-            if let Some(cn) = class_name {
-                let resolved = loan_env!(self, resolve_private_method_for_vm(&cn, method, &args));
-                if let Some((owner_class, method_def)) = resolved {
-                    let caller_allowed = self.can_fast_dispatch_private_method_vm(&owner_class);
-                    if caller_allowed && let Some(ref cc) = method_def.compiled_code {
-                        let cc = cc.clone();
-                        let target_id = match &target {
-                            Value::Instance { id, .. } => Some(*id),
-                            _ => None,
-                        };
-                        let attrs_cell = match &target {
-                            Value::Instance { attributes, .. } => Some(attributes.clone()),
-                            _ => None,
-                        };
-                        let attributes = match &target {
-                            Value::Instance { attributes, .. } => attributes.to_map(),
-                            _ => std::collections::HashMap::new(),
-                        };
-                        let invocant_for_dispatch = if attributes.is_empty() {
-                            Value::Package(crate::symbol::Symbol::intern(&cn))
-                        } else {
-                            target.clone()
-                        };
-                        let pushed_dispatch = loan_env!(
-                            self,
-                            push_method_dispatch_frame(&cn, method, &args, invocant_for_dispatch,)
-                        );
-                        let invocant = Some(target);
-                        let empty_fns = HashMap::new();
-                        let method_result = self.call_compiled_method(
-                            &cn,
-                            &owner_class,
-                            method,
-                            &method_def,
-                            &cc,
-                            attributes,
-                            args,
-                            invocant,
-                            &empty_fns,
-                        );
-                        if pushed_dispatch {
-                            self.pop_method_dispatch();
-                        }
-                        self.pop_method_samewith_context();
-                        let (result, new_attrs, attrs_adjusted) = method_result?;
-                        if let Some(id) = target_id {
-                            // Commit only a `:=`-adjusted snapshot: an unadjusted
-                            // one equals the cell and the whole-map write would
-                            // race with concurrent cell-CAS (lost updates).
-                            if attrs_adjusted && let Some(cell) = &attrs_cell {
-                                cell.commit_attrs(new_attrs.clone());
-                            }
-                            if !self.in_lvalue_assignment
-                                && let Value::Proxy { ref fetcher, .. } = result
-                            {
-                                // Without a `:=` adjustment the triple's map is
-                                // the stale entry snapshot (the lazy reconcile
-                                // skips the cell clone) — re-snapshot the live
-                                // cell for the proxy fetcher.
-                                let proxy_attrs = if !attrs_adjusted && let Some(cell) = &attrs_cell
-                                {
-                                    cell.to_map()
-                                } else {
-                                    new_attrs
-                                };
-                                return loan_env!(
-                                    self,
-                                    proxy_fetch(fetcher, None, &cn, &proxy_attrs, id)
-                                );
-                            }
-                        }
-                        return Ok(result);
-                    }
-                }
-            }
-        }
-        // User-defined ^method (metamethod) dispatch:
-        // Foo.^bar passes Foo as the first positional argument.
-        if method.starts_with('^')
-            && method.len() > 1
-            && !crate::runtime::Interpreter::is_classhow_method(&method[1..])
-        {
-            let class_name = match &target {
-                Value::Instance { class_name, .. } => Some(class_name.resolve()),
-                Value::Package(name) => Some(name.resolve()),
-                _ => None,
-            };
-            if let Some(cn) = class_name
-                && self.has_user_method(&cn, method)
-            {
-                let mut how_args = vec![target.clone()];
-                how_args.extend(args);
-                // CARRIER: user-defined ^metamethod dispatch (MOP). See ledger §C.
-                crate::vm::vm_stats::record_method_fallback(method);
-                return loan_env!(self, call_method_with_values(target, method, how_args));
-            }
-        }
-        // Only attempt compiled path for Instance or Package targets
-        let class_name = match &target {
-            Value::Instance { class_name, .. } => Some(class_name.resolve()),
-            Value::Package(name) => Some(name.resolve()),
-            _ => None,
-        };
-        if let Some(ref cn) = class_name {
-            let class_sym = crate::symbol::Symbol::intern(cn);
-            let method_sym = crate::symbol::Symbol::intern(method);
-            let cache_key = (class_sym, method_sym);
-
-            // Fast method dispatch cache: skip wrap chain check, compiled_code
-            // extraction, and param_def eligibility scans for known-fast methods.
-            if !self.has_any_wrap_chains()
-                && let Some(entry) = self.fast_method_cache.get(&cache_key)
-                && args.len() <= entry.positional_count
-            {
-                let needs_default_eval =
-                    entry.has_defaults && args.len() < entry.positional_count && {
-                        let mut pos = 0;
-                        entry.method_def.param_defs.iter().any(|pd| {
-                            if pd.is_invocant || pd.traits.iter().any(|t| t == "invocant") {
-                                return false;
-                            }
-                            let result = pos >= args.len() && pd.default.is_some();
-                            pos += 1;
-                            result
-                        })
-                    };
-                let has_attr_aliases = match &target {
-                    Value::Instance { attributes, .. } => attributes
-                        .as_map()
-                        .keys()
-                        .any(|k| k.starts_with(super::vm_method_dispatch::ATTR_ALIAS_META_PREFIX)),
-                    _ => false,
-                };
-                // Slice 2d (method follow-up): a container variable passed into a
-                // plain scalar `$` param must share the caller's container, which
-                // only the slow `bind_function_args_values` path realizes. Skip
-                // the cached fast path so it falls through to the full resolve
-                // (which dispatches with `None` -> `call_compiled_method`).
-                let shares_scalar_container =
-                    self.method_shares_container_into_scalar_param(&entry.method_def, &args);
-                if !needs_default_eval && !has_attr_aliases && !shares_scalar_container {
-                    let owner_class = entry.owner_class.resolve();
-                    let method_def = entry.method_def.clone();
-                    let cc = entry.compiled_code.clone();
-                    let can_skip_merge = entry.can_skip_merge;
-                    return self.dispatch_compiled_method(
-                        cn,
-                        &owner_class,
-                        method,
-                        &method_def,
-                        &cc,
-                        target,
-                        args,
-                        Some(can_skip_merge),
-                    );
-                }
-            }
-
-            if let Some((owner_class, method_def)) = {
-                // Monomorphic inline cache: single-entry check before HashMap.
-                if let Some((cc, cm, ref co, ref cd)) = self.last_method_resolve
-                    && cc == class_sym
-                    && cm == method_sym
-                    && !cd.is_multi
-                {
-                    Some((co.clone(), cd.clone()))
-                } else {
-                    let cached = self.method_resolve_cache.get(&cache_key).cloned();
-                    let result: Option<(String, std::sync::Arc<crate::runtime::MethodDef>)> =
-                        if let Some(ref hit) = cached
-                            && let Some((_, def)) = hit
-                            && !def.is_multi
-                        {
-                            hit.clone()
-                        } else if let Some(arg_keys) = Self::multi_arg_type_keys(&args)
-                            && self.multi_dispatch_type_cacheable(class_sym, method_sym, cn, method)
-                        {
-                            // Sound multi-method resolution cache (§B): a type+arity-
-                            // deterministic multi resolves as a function of (class,
-                            // method, positional arg types), so key it on those types
-                            // rather than re-running the MRO/specificity walk per call.
-                            let mkey = (class_sym, method_sym, arg_keys);
-                            if let Some(hit) = self.multi_resolve_cache.get(&mkey) {
-                                hit.clone()
-                            } else {
-                                let resolved = loan_env!(
-                                    self,
-                                    resolve_method_with_owner_invocant(cn, method, &args, &target)
-                                );
-                                let resolved_arc =
-                                    resolved.map(|(owner, def)| (owner, std::sync::Arc::new(def)));
-                                // Never cache an ambiguous multi resolution — it must
-                                // re-raise X::Multi::Ambiguous on every call.
-                                if !self.dispatch_ambiguous {
-                                    self.multi_resolve_cache.insert(mkey, resolved_arc.clone());
-                                }
-                                resolved_arc
-                            }
-                        } else {
-                            let resolved = loan_env!(
-                                self,
-                                resolve_method_with_owner_invocant(cn, method, &args, &target)
-                            );
-                            let resolved_arc =
-                                resolved.map(|(owner, def)| (owner, std::sync::Arc::new(def)));
-                            if resolved_arc.as_ref().is_none_or(|(_, def)| !def.is_multi) {
-                                self.method_resolve_cache
-                                    .insert(cache_key, resolved_arc.clone());
-                            }
-                            resolved_arc
-                        };
-                    if let Some((ref owner, ref def)) = result
-                        && !def.is_multi
-                    {
-                        self.last_method_resolve =
-                            Some((class_sym, method_sym, owner.clone(), def.clone()));
-                    }
-                    result
-                }
-            } {
-                if let Some(result) = self.check_method_wrap_chain(
-                    cn,
-                    &owner_class,
-                    method,
-                    &method_def,
-                    &target,
-                    &args,
-                ) {
-                    return result;
-                }
-                // Resolve to a method def that has compiled bytecode. Normally
-                // `compiled_code` is already populated at class registration. A
-                // few sites add methods after that compile pass (e.g.
-                // `.^add_method`, ledger §1) and leave `compiled_code = None`;
-                // for those, compile on demand here so the method runs as
-                // bytecode instead of tree-walking through the interpreter.
-                let compiled_def: Option<(String, std::sync::Arc<crate::runtime::MethodDef>)> =
-                    if method_def.compiled_code.is_some() {
-                        Some((owner_class.clone(), method_def.clone()))
-                    } else if !method_def.body.is_empty()
-                        && let Some((owner, def)) = self.populate_uncompiled_method(
-                            cn,
-                            &owner_class,
-                            method,
-                            &args,
-                            &target,
-                        )
-                    {
-                        // Refresh the resolve caches so future calls take the
-                        // already-compiled fast path without re-resolving.
-                        self.method_resolve_cache
-                            .insert(cache_key, Some((owner.clone(), def.clone())));
-                        if !def.is_multi {
-                            self.last_method_resolve =
-                                Some((class_sym, method_sym, owner.clone(), def.clone()));
-                        }
-                        Some((owner, def))
-                    } else {
-                        None
-                    };
-                if let Some((owner_class, method_def)) = compiled_def {
-                    let cc = method_def.compiled_code.clone().expect("compiled_code set");
-
-                    // Try to populate fast_method_cache for future calls
-                    if !method_def.is_multi {
-                        self.try_populate_fast_cache(cache_key, cn, &owner_class, &method_def, &cc);
-                    }
-
-                    return self.dispatch_compiled_method(
-                        cn,
-                        &owner_class,
-                        method,
-                        &method_def,
-                        &cc,
-                        target,
-                        args,
-                        None,
-                    );
-                }
-            }
-        }
-        // Native `.map` / `.grep` over a concrete array with a simple block: run
-        // the iteration loop in the Interpreter instead of the interpreter (lever A). No
-        // `target_name` here (the receiver is a value, not a mutable variable),
-        // so `$_`-mutating blocks fall back to the interpreter.
-        if let Some(result) = self.try_native_array_map(None, &target, method, &args) {
-            return result;
-        }
-        // Native `.subst` over a Str with a simple pattern/replacement (lever A).
-        if let Some(result) = self.try_native_subst(&target, method, &args) {
-            return result;
-        }
-        // Native `.sort` over a plain array with no/simple comparator (lever A).
-        if let Some(result) = self.try_native_sort(&target, method, &args) {
-            return result;
-        }
-        // Native `.min` / `.max` over a plain list, including `:by` blocks.
-        if let Some(result) = self.try_native_extrema(&target, method, &args) {
-            return result;
-        }
-        // Native `.minmax` over a plain list, including `:by` blocks.
-        if let Some(result) = self.try_native_minmax(&target, method, &args) {
-            return result;
-        }
-        // Native `.first` over a plain list (no-adverb forms), including blocks.
-        if let Some(result) = self.try_native_first(&target, method, &args) {
-            return result;
-        }
-        // Native QuantHash coercion `.Set`/`.Bag`/`.Mix`/`.SetHash`/`.BagHash`/
-        // `.MixHash` over a list-like aggregate or a plain Cool scalar (ledger §D:
-        // native receiver dispatch -> Interpreter-native). Pure element-folding
-        // shared with the interpreter via `builtins::quanthash_coerce`. `.MixHash`
-        // embeds its type metadata directly in the `Value::Mix` Arc (not an
-        // interpreter-owned side table — container metadata has travelled in the
-        // value since #2952), so it is a pure value op like the others. A scalar
-        // (`42.Set`) becomes a single-element collection. Instance/Package/Junction
-        // receivers fall through.
-        if args.is_empty()
-            && let Some(result) = Self::try_native_quanthash_coerce(&target, method)
-        {
-            return result;
-        }
-        // Native `.Map` / `.Hash` coercion over a list-like aggregate or a plain
-        // Cool scalar (pure value op; the `Map` declared-type is embedded in the
-        // Hash Arc). A scalar (`42.Hash`) raises the same odd-number error as the
-        // interpreter. Instance/Package/Junction fall through.
-        if args.is_empty()
-            && let Some(result) = Self::try_native_map_hash_coerce(&target, method)
-        {
-            return result;
-        }
-        // Native `.Seq` coercion over a structural receiver (Seq/Array/Slip/
-        // Range/bare scalar). Supply/LazyList/Instance need state and fall
-        // through to the interpreter.
-        if args.is_empty()
-            && method == "Seq"
-            && let Some(result) = crate::builtins::seq_coerce::to_seq_structural(&target)
-        {
-            return Ok(result);
-        }
-        // Native `.iterator` construction over a plain receiver (Range/Set/Bag/
-        // Mix/List/Array/...): builds the `Iterator` instance via the single
-        // `builtins::iterator_construct` impl the interpreter also uses. `Seq`
-        // (consumed-state + `squish` env mutation) and an already-built Iterator
-        // fall through to the interpreter.
-        if args.is_empty()
-            && let Some(result) = Self::try_native_iterator_construct(&target, method)
-        {
-            return Ok(result);
-        }
-        // Native `.IO` coercion over a Cool scalar (`"path".IO`, `42.IO`) — builds
-        // an IO::Path via the shared `make_io_path_instance` (ledger §D). Instance /
-        // non-IO Package / aggregate receivers fall through.
-        if let Some(result) = self.try_native_io_coercion(&target, method, &args) {
-            return result;
-        }
-        // Native `.encode` (Cool scalar -> Buf) / `.decode` (Buf/Blob -> Str) —
-        // pure transformation via the VM-owned encoding registry, no `io_handles`
-        // (ledger §D). Single impl shared with the interpreter catch-all.
-        if let Some(result) = self.try_native_encode_decode(&target, method, &args) {
-            return result;
-        }
-        // TODO: compile to bytecode — native/Buf/Failure method fork (ledger §1).
-        // User-defined Instance methods now always run as bytecode (compiled at
-        // registration, or on demand above via `populate_uncompiled_method`), so
-        // what remains here is native receiver dispatch (Buf/Blob/Failure/IO and
-        // the `dispatch_method_by_name_*` machinery) that depends on interpreter-
-        // owned state (③ state ownership / first-class container Phase 2).
-        crate::vm::vm_stats::record_method_fallback(method);
-        self.vm_call_method_with_values(target, method, args)
-    }
-
-    /// Interpreter-native dispatch for the pure-handle methods of an `IO::Handle`
-    /// (`close`/`tell`/`eof`/`seek`/`opened`/`t` plus the PR-D Tier-1
-    /// setters/getters `chomp`/`nl-out`/`out-buffer`/`encoding` and
-    /// `native-descriptor`) — the ones that touch only the handle's own state,
-    /// with no `emit_output` / env / encoding-helper dependency.
-    /// Operates on the Interpreter's own [`io_handles`](Interpreter::io_handles_mut) handle and the
-    /// shared `IoHandleState` methods (the single authoritative impl the
-    /// interpreter's `*_handle_value` wrappers also use), so behavior is identical
-    /// to the interpreter's native fork.
-    ///
-    /// Returns `None` (fall through to the interpreter) for any other receiver,
-    /// any other method, unexpected arity, or junction arguments (which need
-    /// interpreter autothreading). Restricted to the exact `"IO::Handle"` class:
-    /// `IO::Socket::INET` (socket-semantic close) and `IO::Pipe` (process-reaping
-    /// close) are intentionally excluded.
-    fn try_native_io_handle_method(
-        &mut self,
-        target: &Value,
-        method: &str,
-        args: &[Value],
-    ) -> Option<Result<Value, RuntimeError>> {
-        let id = match target {
-            Value::Instance {
-                class_name,
-                attributes,
-                ..
-            } if class_name == "IO::Handle" => match attributes.as_map().get("handle") {
-                Some(Value::Int(i)) if *i >= 0 => *i as usize,
-                _ => return None,
-            },
-            _ => return None,
-        };
-
-        // Pure-handle setters/getters whose argument touches only handle state
-        // (PLAN.md ③ native IO PR-D Tier-1). Junction arguments fall through to
-        // the interpreter for autothreading.
-        if args.iter().any(|a| matches!(a, Value::Junction { .. })) {
-            return None;
-        }
-
-        // `.flush` is target-agnostic and pure (flush pending out-buffer + the OS
-        // file buffer); shared `flush_for_method` (PR-D Tier-2b). A handle id not
-        // in the table falls through — the interpreter shapes that into an
-        // `X::IO::Flush` Failure (a value, not an error), which this Err-on-absent
-        // path below cannot reproduce.
-        if method == "flush" && args.is_empty() {
-            let mut table = self.io_handles_mut();
-            return table
-                .map
-                .get_mut(&id)
-                .map(|state| state.flush_for_method().map(|_| Value::Bool(true)));
-        }
-
-        // `.encoding` returns Nil for binary mode and a Str otherwise, so its
-        // result shaping is method-specific (mirrors the interpreter's
-        // `native_io` arm exactly).
-        enum EncodingOp {
-            /// Getter: shape the current encoding into Nil("bin") / Str.
-            Get,
-            /// Setter to binary mode (`:bin`, `"bin"`, or `Nil` arg) -> Nil.
-            SetBin,
-            /// Setter to a named encoding -> Str(encoding).
-            Set(String),
-        }
-
-        enum Op {
-            Tell,
-            Eof,
-            Opened,
-            Tty,
-            Close,
-            Seek(i64, i32),
-            Chomp(Option<bool>),
-            NlOut(Option<String>),
-            OutBuffer(Option<Option<usize>>),
-            Encoding(EncodingOp),
-            NativeDescriptor,
-        }
-        let op = match method {
-            "tell" if args.is_empty() => Op::Tell,
-            "eof" if args.is_empty() => Op::Eof,
-            "opened" if args.is_empty() => Op::Opened,
-            "t" if args.is_empty() => Op::Tty,
-            "close" if args.is_empty() => Op::Close,
-            "native-descriptor" if args.is_empty() => Op::NativeDescriptor,
-            "chomp" => Op::Chomp(args.first().map(|a| a.truthy())),
-            "nl-out" => Op::NlOut(args.first().map(|a| a.to_string_value())),
-            "out-buffer" => Op::OutBuffer(
-                args.first()
-                    .map(crate::runtime::Interpreter::parse_out_buffer_size),
-            ),
-            "encoding" => Op::Encoding(match args.first() {
-                None => EncodingOp::Get,
-                Some(Value::Nil) => EncodingOp::SetBin,
-                Some(arg) => {
-                    let enc = arg.to_string_value();
-                    if enc == "bin" {
-                        EncodingOp::SetBin
-                    } else {
-                        EncodingOp::Set(enc)
-                    }
-                }
-            }),
-            "seek" => {
-                // seek($offset, $whence = SeekFromBeginning). Mirror the
-                // interpreter's `native_io_handle` arg handling exactly: a
-                // non-Int offset coerces to 0 (`unwrap_or(0)`), and the whence
-                // string maps to 0/1/2 (anything else -> 0). Defer to the
-                // interpreter when an arg is a junction (needs autothreading) or
-                // the arity is unexpected.
-                if args.len() > 2 || args.iter().any(|a| matches!(a, Value::Junction { .. })) {
-                    return None;
-                }
-                let pos = match args.first() {
-                    Some(Value::Int(i)) => *i,
-                    _ => 0,
-                };
-                let mode = match args.get(1) {
-                    Some(v) => match v.to_string_value().as_str() {
-                        "SeekFromCurrent" => 1,
-                        "SeekFromEnd" => 2,
-                        _ => 0,
-                    },
-                    None => 0,
-                };
-                Op::Seek(pos, mode)
-            }
-            _ => return None,
-        };
-
-        let mut table = self.io_handles_mut();
-        let Some(state) = table.map.get_mut(&id) else {
-            return Some(Err(RuntimeError::new("Invalid IO::Handle")));
-        };
-        let result = match op {
-            Op::Tell => state.tell().map(Value::Int),
-            Op::Eof => state.eof().map(Value::Bool),
-            Op::Opened => Ok(Value::Bool(state.is_opened())),
-            Op::Tty => Ok(Value::Bool(state.is_tty())),
-            Op::Close => state.close().map(Value::Bool),
-            Op::Seek(pos, mode) => state.seek(pos, mode).map(Value::Int),
-            Op::Chomp(set) => Ok(Value::Bool(state.chomp_setting(set))),
-            Op::NlOut(set) => Ok(Value::str(state.nl_out_setting(set))),
-            Op::OutBuffer(set) => state.out_buffer_setting(set).map(|n| Value::Int(n as i64)),
-            Op::NativeDescriptor => state.native_descriptor().map(Value::Int),
-            Op::Encoding(enc_op) => Ok(match enc_op {
-                EncodingOp::Get => {
-                    let cur = state.encoding_setting(None);
-                    if cur == "bin" {
-                        Value::Nil
-                    } else {
-                        Value::str(cur)
-                    }
-                }
-                EncodingOp::SetBin => {
-                    state.encoding_setting(Some("bin".to_string()));
-                    Value::Nil
-                }
-                EncodingOp::Set(enc) => {
-                    state.encoding_setting(Some(enc.clone()));
-                    Value::str(enc)
-                }
-            }),
-        };
-        Some(result)
-    }
-
-    /// Interpreter-native text output for a `File`+UTF8 `IO::Handle` receiver
-    /// (`print`/`put`/`say`/`printf`/`print-nl`): build the payload exactly as
-    /// the interpreter's `native_io` handlers do — `render_str_value` for
-    /// print/put, `render_gist_value` for say (byte-identical; the same helpers
-    /// the Interpreter's own `say`/`print` ops use), the pure `sprintf` helpers for
-    /// printf — and write it through the Interpreter's `io_handles` handle via the shared
-    /// `IoHandleState::native_text_write` (PLAN.md ③ native IO PR-D Tier-2a/2b).
-    ///
-    /// Returns `None` (fall through to the interpreter) for any non-`IO::Handle`
-    /// receiver, any other method, a junction argument (autothreading), or a
-    /// handle whose target/encoding is not File+UTF8 — Stdout/Stderr need
-    /// `emit_output`/`stderr_output` and a non-UTF8 File needs
-    /// `encode_with_encoding`, neither Interpreter-reachable yet (③ 後段/④). The
-    /// target/encoding gate is read *before* the payload is built so a
-    /// fall-through never double-runs the argument stringification.
-    fn try_native_io_handle_output(
-        &mut self,
-        target: &Value,
-        method: &str,
-        args: &[Value],
-    ) -> Option<Result<Value, RuntimeError>> {
-        let id = match target {
-            Value::Instance {
-                class_name,
-                attributes,
-                ..
-            } if class_name == "IO::Handle" => match attributes.as_map().get("handle") {
-                Some(Value::Int(i)) if *i >= 0 => *i as usize,
-                _ => return None,
-            },
-            _ => return None,
-        };
-
-        enum Kind {
-            Print,
-            Put,
-            Say,
-            Printf,
-            PrintNl,
-        }
-        let kind = match method {
-            "print" => Kind::Print,
-            "put" => Kind::Put,
-            "say" => Kind::Say,
-            "printf" => Kind::Printf,
-            "print-nl" => Kind::PrintNl,
-            _ => return None,
-        };
-        // Junction args autothread in the interpreter; fall through for those
-        // (printf's first-arg junction also threads there).
-        if args.iter().any(|a| matches!(a, Value::Junction { .. })) {
-            return None;
-        }
-
-        // Resolve the target *before* building the payload, so falling through
-        // (Socket / non-UTF8 File / Stdin) never runs the arg stringification
-        // twice. Stdout/Stderr emit via the Interpreter's shared output sink (③後段 PR-C);
-        // File writes its handle state (Tier-2a).
-        enum Tgt {
-            File,
-            Stdout,
-            Stderr,
-        }
-        let tgt = {
-            let table = self.io_handles_mut();
-            let state = table.map.get(&id)?;
-            if state.can_native_text_write() {
-                Tgt::File
-            } else if state.is_stdout_target() {
-                Tgt::Stdout
-            } else if state.is_stderr_target() {
-                Tgt::Stderr
-            } else {
-                return None;
-            }
-        };
-
-        // Build the argument content. `print-nl` has no args — its payload is the
-        // handle's `nl_out`, which `prepare_text_payload` appends via `newline`
-        // (content = "", newline = true). Its closed-handle error uses "write",
-        // matching the interpreter's `print-nl` (which calls `write_to_handle_value`).
-        let (content, newline, trying): (String, bool, &str) = match kind {
-            Kind::PrintNl => (String::new(), true, "write"),
-            // printf: validate the directives then format, exactly as the
-            // interpreter's `printf` arm (pure `sprintf` helpers, no handle state).
-            Kind::Printf => {
-                let fmt = args
-                    .first()
-                    .map(|v| v.to_string_value())
-                    .unwrap_or_default();
-                let rest = if args.is_empty() { &[][..] } else { &args[1..] };
-                if let Err(e) =
-                    crate::runtime::sprintf::validate_sprintf_directives(&fmt, rest.len())
-                {
-                    return Some(Err(e));
-                }
-                (
-                    crate::runtime::sprintf::format_sprintf_args(&fmt, rest),
-                    false,
-                    method,
-                )
-            }
-            Kind::Say => {
-                let mut c = String::new();
-                for arg in args {
-                    c.push_str(&loan_env!(self, render_gist_value(arg)));
-                }
-                (c, true, method)
-            }
-            // print / put use `render_str_value`.
-            Kind::Print => {
-                let mut c = String::new();
-                for arg in args {
-                    c.push_str(&loan_env!(self, render_str_value(arg)));
-                }
-                (c, false, method)
-            }
-            Kind::Put => {
-                let mut c = String::new();
-                for arg in args {
-                    c.push_str(&loan_env!(self, render_str_value(arg)));
-                }
-                (c, true, method)
-            }
-        };
-
-        // File: prepare + buffered file write in one confined guard.
-        if matches!(tgt, Tgt::File) {
-            let mut table = self.io_handles_mut();
-            let Some(state) = table.map.get_mut(&id) else {
-                return Some(Err(RuntimeError::new("Invalid IO::Handle")));
-            };
-            return Some(
-                state
-                    .native_text_write(&content, newline, trying)
-                    .map(|_| Value::Bool(true)),
-            );
-        }
-
-        // Stdout/Stderr: prepare the payload on the receiver handle (closed check
-        // + nl_out + bytes_written), then emit via the shared output sink.
-        let payload = {
-            let mut table = self.io_handles_mut();
-            let Some(state) = table.map.get_mut(&id) else {
-                return Some(Err(RuntimeError::new("Invalid IO::Handle")));
-            };
-            match state.prepare_text_payload(&content, newline, trying) {
-                Ok(p) => p,
-                Err(e) => return Some(Err(e)),
-            }
-        };
-        if matches!(tgt, Tgt::Stdout) {
-            self.vm_emit_stdout(&payload);
-        } else {
-            self.vm_emit_stderr(&payload);
-        }
-        Some(Ok(Value::Bool(true)))
-    }
-
-    /// Interpreter-native raw byte output for a `File` `IO::Handle` receiver
-    /// (`write` / `spurt`): build the bytes exactly as the interpreter's
-    /// `native_io` handlers do and write them straight to the file via the
-    /// shared `IoHandleState::native_write_bytes_file` (raw, `:out-buffer`- and
-    /// encoding-bypassing — same semantics as `write_bytes_to_handle_value`).
-    /// PLAN.md ③ native IO PR-D Tier-2c.
-    ///
-    /// `write` concatenates each arg's bytes — buffer types (Buf/Blob/utf8/
-    /// utf16) via `supply_chunk_to_bytes` (utf16-aware), non-buffers via
-    /// `render_str_value` (their UTF-8 bytes). `spurt` writes its single
-    /// argument: a Buf's raw bytes, or a Str's UTF-8 bytes.
-    ///
-    /// Returns `None` (fall through) for any non-`IO::Handle` receiver, any other
-    /// method, a junction argument, a non-File target (Stdout/Stderr need
-    /// `emit_output`), or a `spurt` of a Str on a non-UTF-8 handle (needs
-    /// `encode_with_encoding`). The File gate is read before the bytes are built.
-    fn try_native_io_handle_byte_output(
-        &mut self,
-        target: &Value,
-        method: &str,
-        args: &[Value],
-    ) -> Option<Result<Value, RuntimeError>> {
-        let id = match target {
-            Value::Instance {
-                class_name,
-                attributes,
-                ..
-            } if class_name == "IO::Handle" => match attributes.as_map().get("handle") {
-                Some(Value::Int(i)) if *i >= 0 => *i as usize,
-                _ => return None,
-            },
-            _ => return None,
-        };
-        if !matches!(method, "write" | "spurt") {
-            return None;
-        }
-        // Junction args autothread in the interpreter; fall through for those.
-        if args.iter().any(|a| matches!(a, Value::Junction { .. })) {
-            return None;
-        }
-
-        // File-target gate, read before building the bytes. For `spurt` of a Str
-        // the handle encoding must also be UTF-8 (a non-UTF-8 encoding re-enters
-        // `encode_with_encoding`); `write` and a Buf `spurt` ignore encoding.
-        let spurt_str_needs_utf8 =
-            method == "spurt" && !args.first().map(Self::is_buf_value).unwrap_or(false);
-        {
-            let table = self.io_handles_mut();
-            let state = table.map.get(&id)?;
-            if !state.is_file_target() {
-                return None;
-            }
-            if spurt_str_needs_utf8 && !state.can_native_text_write() {
-                return None;
-            }
-        }
-
-        let bytes: Vec<u8> = if method == "spurt" {
-            let content_value = args
-                .first()
-                .cloned()
-                .unwrap_or_else(|| Value::Str(String::new().into()));
-            if Self::is_buf_value(&content_value) {
-                Self::extract_buf_bytes(&content_value)
-            } else {
-                // Gate guaranteed UTF-8, so the Str's bytes are its UTF-8 bytes.
-                content_value.to_string_value().into_bytes()
-            }
-        } else {
-            // write: concatenate each arg — buffer types via supply_chunk_to_bytes
-            // (utf16-aware), non-buffers via render_str_value (UTF-8 bytes).
-            let mut out = Vec::new();
-            for arg in args {
-                if Self::is_buf_value(arg) {
-                    out.extend(self.supply_chunk_to_bytes(arg, "utf-8"));
-                } else {
-                    out.extend(loan_env!(self, render_str_value(arg)).into_bytes());
-                }
-            }
-            out
-        };
-
-        let mut table = self.io_handles_mut();
-        let Some(state) = table.map.get_mut(&id) else {
-            return Some(Err(RuntimeError::new("Invalid IO::Handle")));
-        };
-        Some(
-            state
-                .native_write_bytes_file(&bytes)
-                .map(|_| Value::Bool(true)),
-        )
-    }
-
-    /// Interpreter-native reads from a `File`+UTF8 `IO::Handle` receiver (③後段 PR-D read
-    /// side): `get` (next line → `Str`/`Nil`), `slurp` (rest of file → `Str`),
-    /// `read` (up to N bytes → `Buf`). Each delegates to the shared
-    /// `IoHandleState` reader and shapes the result exactly as the interpreter's
-    /// handler.
-    ///
-    /// Returns `None` (fall through) for any non-`IO::Handle` receiver, any other
-    /// method, or a handle the native path can't serve — Stdin / ArgFiles (need
-    /// `@*ARGS`), a non-UTF-8 / `:bin` slurp (needs `Buf` / `decode_with_encoding`),
-    /// junction args — keeping the interpreter's richer read.
-    fn try_native_io_handle_read(
-        &mut self,
-        target: &Value,
-        method: &str,
-        args: &[Value],
-    ) -> Option<Result<Value, RuntimeError>> {
-        let id = match target {
-            Value::Instance {
-                class_name,
-                attributes,
-                ..
-            } if class_name == "IO::Handle" => match attributes.as_map().get("handle") {
-                Some(Value::Int(i)) if *i >= 0 => *i as usize,
-                _ => return None,
-            },
-            _ => return None,
-        };
-        if args.iter().any(|a| matches!(a, Value::Junction { .. })) {
-            return None;
-        }
-        match method {
-            "get" => {
-                if !args.is_empty() {
-                    return None;
-                }
-                let mut table = self.io_handles_mut();
-                let state = table.map.get_mut(&id)?;
-                // File + UTF-8/binary only; everything else falls through.
-                if !state.can_native_text_write() {
-                    return None;
-                }
-                Some(
-                    state
-                        .read_line_native()
-                        .map(|line| line.map(Value::str).unwrap_or(Value::Nil)),
-                )
-            }
-            "slurp" => {
-                // `:bin` falls through (returns a Buf via the interpreter).
-                let has_bin = args
-                    .iter()
-                    .any(|a| matches!(a, Value::Pair(k, v) if k == "bin" && v.truthy()));
-                let mut table = self.io_handles_mut();
-                let state = table.map.get_mut(&id)?;
-                if !state.can_native_slurp_string(has_bin) {
-                    return None;
-                }
-                Some(state.slurp_string_native().map(Value::str))
-            }
-            "read" => {
-                // `.read` returns a Buf. Parse the byte count exactly as the
-                // interpreter (positive Int → that many bytes; else read to EOF).
-                let count = match args.first() {
-                    Some(Value::Int(i)) if *i > 0 => *i as usize,
-                    None => 0,
-                    // Any other first arg (0/negative Int, non-Int) → the
-                    // interpreter coerces to 0 (read all); match that.
-                    Some(Value::Int(_)) => 0,
-                    _ => return None,
-                };
-                let mut table = self.io_handles_mut();
-                let state = table.map.get_mut(&id)?;
-                if !state.is_file_target() {
-                    return None;
-                }
-                Some(
-                    state
-                        .read_bytes_native(count)
-                        .map(crate::runtime::Interpreter::make_buf),
-                )
-            }
-            "getc" => {
-                if !args.is_empty() {
-                    return None;
-                }
-                let mut table = self.io_handles_mut();
-                let state = table.map.get_mut(&id)?;
-                // File + UTF-8/binary only (utf16/non-UTF8 need the interpreter).
-                if !state.can_native_text_write() {
-                    return None;
-                }
-                // One character, possibly multi-byte; empty -> Nil (as `getc`).
-                Some(state.read_chars_native(Some(1)).map(|s| {
-                    if s.is_empty() {
-                        Value::Nil
-                    } else {
-                        Value::str(s)
-                    }
-                }))
-            }
-            "readchars" => {
-                let mut table = self.io_handles_mut();
-                let state = table.map.get_mut(&id)?;
-                if !state.can_native_text_write() {
-                    return None;
-                }
-                // count: a non-negative integer (parsed as the interpreter does);
-                // an invalid arg is the same error; no arg reads to EOF.
-                let count = match args.first() {
-                    None => None,
-                    Some(arg) => match crate::runtime::Interpreter::parse_out_buffer_size(arg) {
-                        Some(n) => Some(n),
-                        None => {
-                            return Some(Err(RuntimeError::new(
-                                "readchars count must be a non-negative integer",
-                            )));
-                        }
-                    },
-                };
-                Some(state.read_chars_native(count).map(Value::str))
-            }
-            _ => None,
-        }
-    }
-
-    /// Whether `target` is a receiver for which the pure `.Set`/`.Bag`/`.Mix`/
-    /// `.Map`/`.Hash`-family coercions can run natively in the VM. Two groups go
-    /// native:
-    ///   * list-like / QuantHash aggregates (List/Array/Seq/Slip/Hash/Set/Bag/Mix/
-    ///     Pair/Range) — they flatten their elements; and
-    ///   * plain `Cool` scalars (Str / Int / BigInt / Num / Rat / FatRat / Complex /
-    ///     Bool) — each becomes a single-element collection via the coercion's
-    ///     catch-all arm. The interpreter routes these exact scalars straight to the
-    ///     same shared `builtins` impl with no special-casing (`Str.Set` →
-    ///     `to_set` `other =>` arm = `Set(blue)`), so going native is behavior-invariant.
-    ///
-    /// `Instance` (`__baggy_data__` / Match captures / user coercion), `Package`
-    /// (type objects — `.Map`/`.Hash` return the type object, handled by the
-    /// interpreter), `Nil`, and `Junction` (autothread) receivers return `false`
-    /// and fall through to the interpreter's richer handling.
-    fn coerce_receiver_native_eligible(target: &Value) -> bool {
-        matches!(
-            target,
-            Value::Array(..)
-                | Value::Seq(_)
-                | Value::Slip(_)
-                | Value::Hash(_)
-                | Value::Set(..)
-                | Value::Bag(..)
-                | Value::Mix(..)
-                | Value::Pair(..)
-                | Value::ValuePair(..)
-                | Value::Str(_)
-                | Value::Int(_)
-                | Value::BigInt(_)
-                | Value::Num(_)
-                | Value::Rat(..)
-                | Value::FatRat(..)
-                | Value::Complex(..)
-                | Value::Bool(_)
-        ) || target.is_range()
-    }
-
-    /// Interpreter-native QuantHash coercion for `.Set`/`.Bag`/`.Mix`/`.SetHash`/
-    /// `.BagHash`/`.MixHash` over a list-like aggregate or a plain `Cool` scalar
-    /// (see [`coerce_receiver_native_eligible`](Self::coerce_receiver_native_eligible)).
-    /// Delegates the element folding to the single authoritative pure
-    /// implementation in `builtins::quanthash_coerce`, the same one the
-    /// interpreter's `dispatch_method_by_name_2` uses, so the result is
-    /// behavior-invariant. The `*Hash` variants flip the mutable flag exactly as
-    /// the interpreter does. `.MixHash` embeds its type metadata directly in the
-    /// `Value::Mix` Arc (container metadata has travelled in the value since #2952),
-    /// so it is a pure value op like the others.
-    ///
-    /// Returns `None` (fall through to the interpreter) for `Instance`/`Package`/
-    /// `Nil`/`Junction` receivers, leaving those rarer cases — `__baggy_data__`
-    /// instances, type objects, user coercion, autothread — to the interpreter.
-    fn try_native_quanthash_coerce(
-        target: &Value,
-        method: &str,
-    ) -> Option<Result<Value, RuntimeError>> {
-        if !matches!(
-            method,
-            "Set" | "SetHash" | "Bag" | "BagHash" | "Mix" | "MixHash"
-        ) {
-            return None;
-        }
-        if !Self::coerce_receiver_native_eligible(target) {
-            return None;
-        }
-        let target = target.clone();
-        let result = match method {
-            "Set" => crate::builtins::quanthash_coerce::to_set(target, "Set"),
-            "SetHash" => {
-                crate::builtins::quanthash_coerce::to_set(target, "SetHash").map(|r| match r {
-                    Value::Set(items, _) => Value::Set(items, true),
-                    other => other,
-                })
-            }
-            "Bag" => crate::builtins::quanthash_coerce::to_bag(target, "Bag"),
-            "BagHash" => {
-                crate::builtins::quanthash_coerce::to_bag(target, "BagHash").map(|r| match r {
-                    Value::Bag(items, _) => Value::Bag(items, true),
-                    other => other,
-                })
-            }
-            "Mix" => crate::builtins::quanthash_coerce::to_mix(target, "Mix"),
-            "MixHash" => crate::builtins::quanthash_coerce::to_mixhash(target),
-            _ => unreachable!(),
-        };
-        Some(result)
-    }
-
-    /// Interpreter-native `.Map` / `.Hash` coercion over a list-like aggregate or
-    /// a plain `Cool` scalar (see
-    /// [`coerce_receiver_native_eligible`](Self::coerce_receiver_native_eligible)),
-    /// sharing the single `builtins::map_hash_coerce` impl the interpreter also
-    /// uses. `.Map` decontainerizes and embeds the `Map` declared-type *in* the
-    /// `Value::Hash` Arc (container metadata travels in the value since #2952 — no
-    /// interpreter-owned side table), so both are pure value ops. A scalar (`42.Hash`)
-    /// raises `X::Hash::Store::OddNumber` via the same `to_hash`, identical to the
-    /// interpreter. Instance (Match / `__baggy_data__`) and Package (type object,
-    /// where `.Map`/`.Hash` returns the type object) receivers fall through to the
-    /// interpreter's richer handling. Lowercase `.hash` is intentionally not handled here.
-    fn try_native_map_hash_coerce(
-        target: &Value,
-        method: &str,
-    ) -> Option<Result<Value, RuntimeError>> {
-        if !matches!(method, "Map" | "Hash") {
-            return None;
-        }
-        if !Self::coerce_receiver_native_eligible(target) {
-            return None;
-        }
-        let target = target.clone();
-        Some(match method {
-            "Map" => crate::builtins::map_hash_coerce::to_map(target),
-            "Hash" => crate::builtins::map_hash_coerce::to_hash(target, true),
-            _ => unreachable!(),
-        })
-    }
-
-    /// Interpreter-native `.IO` coercion over a `Cool` scalar receiver
-    /// (`"path".IO`, `42.IO`): constructs an `IO::Path` from the receiver's string
-    /// value via the single shared `make_io_path_instance` (which inherits `$*SPEC`
-    /// / `$*CWD` from env — a normal native env read, identical to the
-    /// interpreter's `dispatch_method_by_name` `.IO` arm). An `IO::Path`(-subclass)
-    /// *type object* (`Value::Package`) returns itself (`IO::Path === IO::Path.IO`).
-    ///
-    /// Returns `None` (fall through to the interpreter) for `Instance` receivers
-    /// (which may have a user `.IO` or be IO::Path/IO::Handle), non-IO `Package`
-    /// type objects, and aggregate / `Junction` receivers (which autothread or
-    /// have their own semantics) — only plain stringifiable scalars go native.
-    fn try_native_io_coercion(
-        &self,
-        target: &Value,
-        method: &str,
-        args: &[Value],
-    ) -> Option<Result<Value, RuntimeError>> {
-        if method != "IO" || !args.is_empty() {
-            return None;
-        }
-        match target {
-            // `.IO` on an IO::Path (sub)class type object returns the type object
-            // itself, so `IO::Path::Unix === IO::Path::Unix.IO`.
-            Value::Package(name) => {
-                let n = name.resolve();
-                if n == "IO::Path" || n.starts_with("IO::Path::") {
-                    Some(Ok(target.clone()))
-                } else {
-                    None
-                }
-            }
-            // Plain stringifiable scalars coerce their string value to an IO::Path.
-            Value::Str(_)
-            | Value::Int(_)
-            | Value::BigInt(_)
-            | Value::Num(_)
-            | Value::Rat(..)
-            | Value::FatRat(..)
-            | Value::Complex(..)
-            | Value::Bool(_) => {
-                let s = target.to_string_value();
-                if s.contains('\0') {
-                    return Some(Err(RuntimeError::new(
-                        "X::IO::Null: Found null byte in pathname",
-                    )));
-                }
-                Some(Ok(self.make_io_path_instance(&s)))
-            }
-            _ => None,
-        }
-    }
-
-    /// Interpreter-native `.iterator` construction, mirroring
-    /// `Interpreter::dispatch_iterator_method`: an already-built `Iterator`
-    /// instance returns itself; a `Seq` falls through (consumed-state tracking +
-    /// `squish` env mutation are interpreter-owned); any other receiver builds an
-    /// `Iterator` instance via the single `builtins::iterator_construct` impl.
-    /// Returns `None` to fall through to the interpreter.
-    fn try_native_iterator_construct(target: &Value, method: &str) -> Option<Value> {
-        if method != "iterator" {
-            return None;
-        }
-        // Already an Iterator instance: identity (interpreter's early return).
-        if let Value::Instance { class_name, .. } = target
-            && class_name == "Iterator"
-        {
-            return Some(target.clone());
-        }
-        // Seq: consumed-state + `squish` env mutation are interpreter-owned.
-        if matches!(target, Value::Seq(_)) {
-            return None;
-        }
-        Some(crate::builtins::iterator_construct::build_iterator_instance(target))
-    }
-
-    /// Interpreter-native Iterator protocol over a self-contained, array-backed `Iterator`
-    /// instance (`items` Array + `index` Int, no `squish_source` callbacks).
-    /// Handles the index-advancing protocol methods `pull-one`/`skip-one`/
-    /// `skip-at-least`/`skip-at-least-pull-one`/`sink-all`, mirroring the
-    /// interpreter's mutating iterator dispatch in `methods_mut.rs` exactly,
-    /// including the identity-based writeback (env + locals) so the receiver
-    /// variable and any aliases observe the advance. Behavior-invariant.
-    ///
-    /// Returns `None` (fall through to the interpreter) for: non-`Iterator`
-    /// receivers; squish iterators (which invoke user `as`/`with` callbacks);
-    /// predictive / coroutine iterators (no concrete `items` array); the
-    /// `push-*` family (it writes pulled elements into an external buffer arg,
-    /// needing array-identity writeback handled by the interpreter); and
-    /// `count-only`/`bool-only` (left to the interpreter's predictive handling).
-    pub(super) fn try_native_iterator(
-        &mut self,
-        target: &Value,
-        method: &str,
-        args: &[Value],
-    ) -> Option<Result<Value, RuntimeError>> {
-        if !matches!(
-            method,
-            "pull-one" | "skip-one" | "skip-at-least" | "skip-at-least-pull-one" | "sink-all"
-        ) {
-            return None;
-        }
-        let Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } = target
-        else {
-            return None;
-        };
-        if class_name.resolve() != "Iterator"
-            || attributes.contains_key("squish_source")
-            || attributes.contains_key("is_lazy")
-        {
-            // squish iterators invoke user callbacks; lazy iterators (gather /
-            // lazy Seq) pull through interpreter-owned coroutine state rather than
-            // a materialized `items` snapshot — leave both to the interpreter.
-            return None;
-        }
-        // Only a concrete array-backed iterator (excludes predictive/coroutine
-        // iterators whose state lives off the instance). Snapshot the (cheap Arc
-        // clone of the) backing array and the start index, then drop the read
-        // guard before the writeback below write-locks the same cell.
-        let (items, start_index) = {
-            let map = attributes.as_map();
-            let Some(Value::Array(items, ..)) = map.get("items") else {
-                return None;
-            };
-            let start_index = match map.get("index") {
-                Some(Value::Int(i)) if *i >= 0 => *i as usize,
-                _ => 0,
-            };
-            (items.clone(), start_index)
-        };
-        let len = items.len();
-        let mut index = start_index;
-        let ret = match method {
-            "pull-one" => {
-                if index < len {
-                    let out = items[index].clone();
-                    index += 1;
-                    out
-                } else {
-                    Value::str_from("IterationEnd")
-                }
-            }
-            "sink-all" => {
-                index = len;
-                Value::str_from("IterationEnd")
-            }
-            "skip-one" => {
-                if index < len {
-                    index += 1;
-                    Value::Bool(true)
-                } else {
-                    Value::Bool(false)
-                }
-            }
-            "skip-at-least" => {
-                let want = args.first().map(crate::runtime::to_int).unwrap_or(0).max(0) as usize;
-                if len.saturating_sub(index) >= want {
-                    index += want;
-                    Value::Bool(true)
-                } else {
-                    index = len;
-                    Value::Bool(false)
-                }
-            }
-            "skip-at-least-pull-one" => {
-                let want = args.first().map(crate::runtime::to_int).unwrap_or(0).max(0) as usize;
-                if len.saturating_sub(index) >= want {
-                    index += want;
-                    if index < len {
-                        let out = items[index].clone();
-                        index += 1;
-                        out
-                    } else {
-                        Value::str_from("IterationEnd")
-                    }
-                } else {
-                    index = len;
-                    Value::str_from("IterationEnd")
-                }
-            }
-            _ => unreachable!(),
-        };
-        // Write the advanced index back by instance identity (env + locals), as
-        // the interpreter's mutating iterator dispatch does, so the receiver
-        // variable and aliases see the advance.
-        if index != start_index {
-            // Write through the shared cell in place: every alias of this
-            // iterator instance (caller var, locals) sees the advance directly.
-            attributes.insert("index".to_string(), Value::Int(index as i64));
-        }
-        Some(Ok(ret))
-    }
-
-    /// Compile a resolved user method's body on demand when it lacks bytecode,
-    /// then dispatch as bytecode instead of through the interpreter bridge.
-    /// Almost all user methods are already compiled at class registration
-    /// (`compile_class_methods`); the gap this closes is methods inserted after
-    /// that pass without their own bytecode — notably `.^add_multi_method`
-    /// (which hardcodes `compiled_code = None`) and any future such site.
-    /// (`.^add_method` already carries the method literal's compiled code.)
-    /// Populates `compiled_code` in the canonical registry (idempotent) and
-    /// re-resolves so the returned def carries the bytecode. Returns `None`
-    /// when the owner is not a user class/role (native receiver) or the body
-    /// stays uncompilable, preserving the interpreter fallback. Ledger §1.
-    fn populate_uncompiled_method(
-        &mut self,
-        cn: &str,
-        owner_class: &str,
-        method: &str,
-        args: &[Value],
-        target: &Value,
-    ) -> Option<(String, std::sync::Arc<crate::runtime::MethodDef>)> {
-        // Both compile passes are no-ops if the name is absent from the
-        // respective registry, so calling both safely covers class- and
-        // role-owned methods. Neither re-enters user code (pure compilation),
-        // so the registry re-entrancy discipline (②) is respected.
-        self.compile_class_methods(owner_class);
-        self.compile_role_methods(owner_class);
-        let (owner, def) = loan_env!(
-            self,
-            resolve_method_with_owner_invocant(cn, method, args, target)
-        )?;
-        if def.compiled_code.is_some() {
-            Some((owner, std::sync::Arc::new(def)))
-        } else {
-            None
-        }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn dispatch_compiled_method(
-        &mut self,
-        cn: &str,
-        owner_class: &str,
-        method: &str,
-        method_def: &std::sync::Arc<crate::runtime::MethodDef>,
-        cc: &std::sync::Arc<CompiledCode>,
-        target: Value,
-        args: Vec<Value>,
-        can_skip_merge: Option<bool>,
-    ) -> Result<Value, RuntimeError> {
-        let target_id = match &target {
-            Value::Instance { id, .. } => Some(*id),
-            _ => None,
-        };
-        let attrs_cell = match &target {
-            Value::Instance { attributes, .. } => Some(attributes.clone()),
-            _ => None,
-        };
-        let attributes = match &target {
-            Value::Instance { attributes, .. } => attributes.to_map(),
-            _ => std::collections::HashMap::new(),
-        };
-        let empty_fns = HashMap::new();
-        let method_result = if let Some(csm) = can_skip_merge {
-            // Fast path: move target directly as base (avoid extra clone).
-            let invocant_for_dispatch = if attributes.is_empty() {
-                Value::Package(crate::symbol::Symbol::intern(cn))
-            } else {
-                target.clone()
-            };
-            let pushed_dispatch = loan_env!(
-                self,
-                push_method_dispatch_frame(cn, method, &args, invocant_for_dispatch,)
-            );
-            let result = self.call_compiled_method_fast(
-                cn,
-                owner_class,
-                method,
-                method_def,
-                cc,
-                attributes,
-                args,
-                target,
-                &empty_fns,
-                csm,
-            );
-            if pushed_dispatch {
-                self.pop_method_dispatch();
-            }
-            self.pop_method_samewith_context();
-            result
-        } else {
-            let invocant_for_dispatch = if attributes.is_empty() {
-                Value::Package(crate::symbol::Symbol::intern(cn))
-            } else {
-                target.clone()
-            };
-            let pushed_dispatch = loan_env!(
-                self,
-                push_method_dispatch_frame(cn, method, &args, invocant_for_dispatch,)
-            );
-            let invocant = Some(target);
-            let result = self.call_compiled_method(
-                cn,
-                owner_class,
-                method,
-                method_def,
-                cc,
-                attributes,
-                args,
-                invocant,
-                &empty_fns,
-            );
-            if pushed_dispatch {
-                self.pop_method_dispatch();
-            }
-            self.pop_method_samewith_context();
-            result
-        };
-        let (result, new_attrs, attrs_adjusted) = method_result?;
-        if let Some(id) = target_id {
-            // Commit only a `:=`-adjusted snapshot: an unadjusted one equals the
-            // cell and the whole-map write would race with concurrent cell-CAS
-            // from another thread (lost updates).
-            if attrs_adjusted && let Some(cell) = &attrs_cell {
-                cell.commit_attrs(new_attrs.clone());
-            }
-            if !self.in_lvalue_assignment
-                && let Value::Proxy { ref fetcher, .. } = result
-            {
-                // Without a `:=` adjustment the triple's map is the stale entry
-                // snapshot (the lazy reconcile skips the cell clone) —
-                // re-snapshot the live cell for the proxy fetcher.
-                let proxy_attrs = if !attrs_adjusted && let Some(cell) = &attrs_cell {
-                    cell.to_map()
-                } else {
-                    new_attrs
-                };
-                return loan_env!(self, proxy_fetch(fetcher, None, cn, &proxy_attrs, id));
-            }
-        }
-        Ok(result)
-    }
-
-    /// Pre-compute and cache fast dispatch eligibility for a method.
-    fn try_populate_fast_cache(
-        &mut self,
-        cache_key: (crate::symbol::Symbol, crate::symbol::Symbol),
-        receiver_class: &str,
-        owner_class: &str,
-        method_def: &std::sync::Arc<crate::runtime::MethodDef>,
-        cc: &std::sync::Arc<CompiledCode>,
-    ) {
-        let has_rw_params = method_def
-            .param_defs
-            .iter()
-            .any(|pd| pd.traits.iter().any(|t| t == "rw"));
-        if has_rw_params {
-            return;
-        }
-        let has_invocant_constraint = method_def.param_defs.iter().any(|pd| {
-            (pd.is_invocant || pd.traits.iter().any(|t| t == "invocant"))
-                && pd.type_constraint.is_some()
-        });
-        let has_complex_params = method_def.param_defs.iter().any(|pd| {
-            if pd.is_invocant || pd.traits.iter().any(|t| t == "invocant") {
-                return false;
-            }
-            if pd.slurpy && pd.name == "%_" {
-                return false;
-            }
-            pd.slurpy
-                || pd.double_slurpy
-                || pd.named
-                || pd.where_constraint.is_some()
-                || pd.sub_signature.is_some()
-                || pd.outer_sub_signature.is_some()
-                || pd.code_signature.is_some()
-                || pd
-                    .type_constraint
-                    .as_ref()
-                    .is_some_and(|tc| tc.contains('('))
-        });
-        let has_role_bindings = self.class_role_param_bindings(owner_class).is_some()
-            || self.class_role_param_bindings(receiver_class).is_some();
-        if has_invocant_constraint || has_complex_params || has_role_bindings {
-            return;
-        }
-        let positional_count = method_def
-            .param_defs
-            .iter()
-            .filter(|pd| {
-                !pd.is_invocant
-                    && !pd.traits.iter().any(|t| t == "invocant")
-                    && !pd.slurpy
-                    && !pd.double_slurpy
-                    && !pd.named
-            })
-            .count();
-        // Also gate on `cc.has_calls`: a body that invokes a CLOSURE
-        // (`$f()` — `CallOnValue`/`CallOnCodeVar`) or a CallDefined/ExecCallSlip can
-        // write a dynamic var / captured-outer lexical / global into this frame's
-        // env (those call ops are NOT in `has_env_writes`). Skipping the merge would
-        // drop that write — e.g. `method go { my $f = { $*x = 1 }; $f() }`. #3658.
-        let can_skip_merge = !cc.has_env_writes && !cc.has_calls;
-        let has_defaults = method_def.param_defs.iter().any(|pd| {
-            !pd.is_invocant && !pd.traits.iter().any(|t| t == "invocant") && pd.default.is_some()
-        });
-        self.fast_method_cache.insert(
-            cache_key,
-            super::FastMethodCacheEntry {
-                owner_class: crate::symbol::Symbol::intern(owner_class),
-                method_def: method_def.clone(),
-                compiled_code: cc.clone(),
-                can_skip_merge,
-                positional_count,
-                has_defaults,
-            },
-        );
-    }
-
-    pub(super) fn try_compiled_method_mut_or_interpret(
-        &mut self,
-        target_name: &str,
-        target: Value,
-        method: &str,
-        args: Vec<Value>,
-    ) -> Result<Value, RuntimeError> {
-        // Native default construction (see `try_compiled_method_or_interpret`).
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && let Some(result) = loan_env!(self, try_native_default_construct(*class_name, &args))
-        {
-            // Pure construction: fresh instance, no caller-env write (Slice 6.3) —
-            // UNLESS the class has a `submethod BUILD`/`TWEAK` whose body can mutate
-            // a captured-outer caller lexical, in which case the dispatch is impure
-            // and the call site must reconcile the caller slot (Slice F twin of the
-            // non-mut path; `reconcile_locals_from_env_at_site`).
-            self.method_dispatch_pure = !self.mro_has_build_or_tweak(&class_name.resolve());
-            return result;
-        }
-        // Native built-in construction (mut path twin of the above).
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && let Some(result) =
-                crate::runtime::Interpreter::try_native_builtin_construct(*class_name, &args)
-        {
-            self.method_dispatch_pure = true;
-            return result;
-        }
-        // Native QuantHash construction (mut path twin of the above).
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && let Some(result) =
-                self.try_native_quanthash_construct_for_package(*class_name, &args)
-        {
-            self.method_dispatch_pure = true;
-            return result;
-        }
-        // Native aggregate construction (mut path twin of the above).
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && let Some(result) =
-                self.try_native_aggregate_construct_for_package(*class_name, &args)
-        {
-            self.method_dispatch_pure = true;
-            return result;
-        }
-        // Native IO::Path family construction (mut path twin of the above).
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && let Some(result) = self.try_native_io_path_construct(*class_name, &args)
-        {
-            self.method_dispatch_pure = true;
-            return result;
-        }
-        // Native Failure construction (mut path twin of the above).
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && class_name.resolve() == "Failure"
-        {
-            self.method_dispatch_pure = true;
-            return Ok(self.build_native_failure_value(&args));
-        }
-        // Native Seq construction (mut path twin of the above).
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && class_name.resolve() == "Seq"
-        {
-            self.method_dispatch_pure = true;
-            return Ok(self.try_native_seq_construct(&args));
-        }
-        // Native IO::Socket::INET construction (mut path twin of the above).
-        if method == "new"
-            && let Value::Package(class_name) = &target
-            && class_name.resolve() == "IO::Socket::INET"
-        {
-            self.method_dispatch_pure = true;
-            return self.dispatch_socket_inet_new(&args);
-        }
-        // Native built-in class method (mut path twin of the above).
-        if let Value::Package(class_name) = &target
-            && let Some(result) = crate::runtime::Interpreter::try_native_builtin_class_method(
-                *class_name,
-                method,
-                &args,
-            )
-        {
-            self.method_dispatch_pure = true;
-            return result;
-        }
-        if let Value::Instance { class_name, .. } = &target {
-            let class = class_name.resolve();
-            // Interpreter-native pure-handle IO dispatch (PLAN.md ③ native IO PR-C/PR-D),
-            // mut path: `$fh.method` on a variable receiver routes here, so the
-            // same state-only `IO::Handle` methods must be intercepted before the
-            // generic native-method fallback below. These methods mutate only the
-            // shared handle-table state (not the receiver binding), so the native
-            // path returns the result directly; `None` falls through unchanged.
-            if let Some(result) = self.try_native_io_handle_method(&target, method, &args) {
-                return result;
-            }
-            // Interpreter-native text output to a File+UTF8 `IO::Handle` (print/put/say/
-            // print-nl), mut path (PR-D Tier-2a). See the non-mut twin above.
-            if let Some(result) = self.try_native_io_handle_output(&target, method, &args) {
-                return result;
-            }
-            // Interpreter-native raw byte output to a File `IO::Handle` (write/spurt):
-            // raw file write, no buffering/encoding (PR-D Tier-2c). Stdout/Stderr
-            // and a non-UTF8 spurt of a Str fall through.
-            if let Some(result) = self.try_native_io_handle_byte_output(&target, method, &args) {
-                return result;
-            }
-            // Interpreter-native line read from a File+UTF8 `IO::Handle` (get): reads via
-            // the handle's record reader (PR-D read side). ArgFiles/Stdin/non-UTF8
-            // (which need @*ARGS / decode) fall through.
-            if let Some(result) = self.try_native_io_handle_read(&target, method, &args) {
-                return result;
-            }
-            // Interpreter-native pure-lexical `IO::Path` methods for variable
-            // receivers (`$p.parent`, `$p.add(...)`): same pure value op as the
-            // non-mut path — produces a *new* IO::Path/string/bool and never mutates
-            // the receiver, so no writeback. Filesystem / cwd-relative forms and
-            // `child :secure` fall through to the native fork below.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) =
-                    Self::try_io_path_lexical(&class, &attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native `.absolute` / `.relative` (path + cwd, lexical — no
-            // filesystem; the VM owns env/cwd). Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) =
-                    self.try_io_path_cwd_method(&attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native filesystem `stat`-only file tests / accessors
-            // (`e`/`f`/`d`/…/`s`/`modified`): resolve the path against the VM-owned
-            // cwd, then `stat` only — no `io_handles`, no content read (ledger §D).
-            // Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) = self.try_io_path_fs_stat(&attributes.as_map(), method)
-            {
-                return result;
-            }
-            // Interpreter-native whole-file content reads (`slurp`/`lines`/`words`):
-            // read the file + split/decode; no `io_handles` (ledger §D). Single impl
-            // shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) =
-                    self.try_io_path_content_read(&attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native single-path filesystem mutations (`spurt`/`mkdir`/
-            // `rmdir`/`unlink`/`chmod`): one-shot syscall, no `io_handles` (ledger
-            // §D). Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) =
-                    self.try_io_path_fs_mutate(&attributes.as_map(), &class, method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native `open`: allocate an `io_handles` entry and return
-            // the `IO::Handle`. The VM owns `io_handles`, so this is a native
-            // dispatch (ledger §D ③). Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) = self.try_io_path_open(&attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native two-path FS ops (`copy`/`rename`/`move`/`symlink`/
-            // `link`): resolve both paths against the VM-owned cwd, one-shot syscall,
-            // no `io_handles` (ledger §D). Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) =
-                    self.try_io_path_two_path_op(&attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // Interpreter-native `comb`: read the file then comb the content (no
-            // `io_handles`; ledger §D). Single impl shared with `native_io_path`.
-            if Self::is_io_path_lexical_class(&class)
-                && let Value::Instance { attributes, .. } = &target
-                && let Some(result) = self.try_io_path_comb(&attributes.as_map(), method, &args)
-            {
-                return result;
-            }
-            // A user-defined subclass of a builtin type may override an inherited
-            // native method (e.g. `class IO::Blob is IO::Handle { method get {…} }`).
-            // The user override must win, so do not take the native fork when the
-            // class (via its MRO) provides its own method of this name.
-            if self.is_native_method(&class, method) && !self.has_user_method(&class, method) {
-                // TODO: compile to bytecode — Instance native-method fork, mut (ledger §1).
-                crate::vm::vm_stats::record_method_fallback(method);
-                return self.vm_call_method_mut_with_values(target_name, target, method, args);
-            }
-        }
-        if matches!(
-            method,
-            "DEFINITE" | "WHAT" | "WHO" | "HOW" | "WHY" | "WHICH" | "WHERE" | "VAR"
-        ) {
-            // CARRIER: MOP pseudo-methods, mut (reflection). See ledger §C.
-            crate::vm::vm_stats::record_method_fallback(method);
-            return loan_env!(
-                self,
-                call_method_mut_with_values(target_name, target, method, args)
-            );
-        }
-        // User-defined ^method (metamethod) dispatch:
-        // Foo.^bar passes Foo as the first positional argument.
-        if method.starts_with('^')
-            && method.len() > 1
-            && !crate::runtime::Interpreter::is_classhow_method(&method[1..])
-        {
-            let class_name = match &target {
-                Value::Instance { class_name, .. } => Some(class_name.resolve()),
-                Value::Package(name) => Some(name.resolve()),
-                _ => None,
-            };
-            if let Some(cn) = class_name
-                && self.has_user_method(&cn, method)
-            {
-                let mut how_args = vec![target.clone()];
-                how_args.extend(args);
-                // CARRIER: user-defined ^metamethod dispatch, mut (MOP). See ledger §C.
-                crate::vm::vm_stats::record_method_fallback(method);
-                return loan_env!(self, call_method_with_values(target, method, how_args));
-            }
-        }
-        if method.starts_with('!') {
-            let class_name = match &target {
-                Value::Instance { class_name, .. } => Some(class_name.resolve()),
-                Value::Package(name) => Some(name.resolve()),
-                _ => None,
-            };
-            if let Some(cn) = class_name {
-                let resolved = loan_env!(self, resolve_private_method_for_vm(&cn, method, &args));
-                if let Some((owner_class, method_def)) = resolved {
-                    let caller_allowed = self.can_fast_dispatch_private_method_vm(&owner_class);
-                    if caller_allowed && let Some(ref cc) = method_def.compiled_code {
-                        let cc = cc.clone();
-                        let target_id = match &target {
-                            Value::Instance { id, .. } => Some(*id),
-                            _ => None,
-                        };
-                        let attrs_cell = match &target {
-                            Value::Instance { attributes, .. } => Some(attributes.clone()),
-                            _ => None,
-                        };
-                        let attributes = match &target {
-                            Value::Instance { attributes, .. } => attributes.to_map(),
-                            _ => std::collections::HashMap::new(),
-                        };
-                        let invocant_for_dispatch = if attributes.is_empty() {
-                            Value::Package(crate::symbol::Symbol::intern(&cn))
-                        } else {
-                            target.clone()
-                        };
-                        let pushed_dispatch = loan_env!(
-                            self,
-                            push_method_dispatch_frame(&cn, method, &args, invocant_for_dispatch,)
-                        );
-                        let invocant = Some(target);
-                        let empty_fns = HashMap::new();
-                        let method_result = self.call_compiled_method(
-                            &cn,
-                            &owner_class,
-                            method,
-                            &method_def,
-                            &cc,
-                            attributes,
-                            args,
-                            invocant,
-                            &empty_fns,
-                        );
-                        if pushed_dispatch {
-                            self.pop_method_dispatch();
-                        }
-                        self.pop_method_samewith_context();
-                        let (result, new_attrs, attrs_adjusted) = method_result?;
-                        if let Some(id) = target_id {
-                            // Commit only a `:=`-adjusted snapshot: an unadjusted
-                            // one equals the cell and the whole-map write would
-                            // race with concurrent cell-CAS (lost updates).
-                            if attrs_adjusted && let Some(cell) = &attrs_cell {
-                                cell.commit_attrs(new_attrs.clone());
-                            }
-                            if !self.in_lvalue_assignment
-                                && let Value::Proxy { ref fetcher, .. } = result
-                            {
-                                // Without a `:=` adjustment the triple's map is
-                                // the stale entry snapshot (the lazy reconcile
-                                // skips the cell clone) — re-snapshot the live
-                                // cell for the proxy fetcher.
-                                let proxy_attrs = if !attrs_adjusted && let Some(cell) = &attrs_cell
-                                {
-                                    cell.to_map()
-                                } else {
-                                    new_attrs
-                                };
-                                return loan_env!(
-                                    self,
-                                    proxy_fetch(fetcher, None, &cn, &proxy_attrs, id)
-                                );
-                            }
-                        }
-                        return Ok(result);
-                    }
-                }
-            }
-        }
-        let class_name = match &target {
-            Value::Instance { class_name, .. } => Some(class_name.resolve()),
-            Value::Package(name) => Some(name.resolve()),
-            _ => None,
-        };
-        if let Some(cn) = class_name
-            && let Some((owner_class, method_def)) = {
-                let class_sym = crate::symbol::Symbol::intern(&cn);
-                let method_sym = crate::symbol::Symbol::intern(method);
-                self.resolve_method_cached(&cn, method, class_sym, method_sym, &args, &target)
-            }
-        {
-            // Ambiguous multi dispatch: two or more candidates matched equally
-            // well. Raise X::Multi::Ambiguous instead of silently picking one.
-            if self.dispatch_ambiguous {
-                self.dispatch_ambiguous = false;
-                let sigs = self.format_method_candidate_signatures(&cn, method);
-                return Err(
-                    crate::runtime::methods_signature_errors::make_multi_ambiguous_error(
-                        method, &cn, &sigs,
-                    ),
-                );
-            }
-            if let Some(result) =
-                self.check_method_wrap_chain(&cn, &owner_class, method, &method_def, &target, &args)
-            {
-                return result;
-            }
-            // Resolve to a def carrying compiled bytecode, compiling on demand
-            // for methods added after the registration compile pass (e.g.
-            // `.^add_method`, ledger §1) so they run as bytecode rather than
-            // tree-walking. None → native receiver, keep interpreter fallback.
-            let resolved: Option<(String, std::sync::Arc<crate::runtime::MethodDef>)> =
-                if method_def.compiled_code.is_some() {
-                    Some((owner_class, method_def))
-                } else if !method_def.body.is_empty() {
-                    self.populate_uncompiled_method(&cn, &owner_class, method, &args, &target)
-                } else {
-                    None
-                };
-            if let Some((owner_class, method_def)) = resolved {
-                let cc = method_def.compiled_code.clone().expect("compiled_code set");
-                let target_id = match &target {
-                    Value::Instance { id, .. } => Some(*id),
-                    _ => None,
-                };
-                let attrs_cell = match &target {
-                    Value::Instance { attributes, .. } => Some(attributes.clone()),
-                    _ => None,
-                };
-                let attributes = match &target {
-                    Value::Instance { attributes, .. } => attributes.to_map(),
-                    _ => std::collections::HashMap::new(),
-                };
-                let invocant_for_dispatch = if attributes.is_empty() {
-                    Value::Package(crate::symbol::Symbol::intern(&cn))
-                } else {
-                    target.clone()
-                };
-                let pushed_dispatch = loan_env!(
-                    self,
-                    push_method_dispatch_frame(&cn, method, &args, invocant_for_dispatch,)
-                );
-                let invocant = Some(target);
-                let empty_fns = HashMap::new();
-                let method_result = self.call_compiled_method(
-                    &cn,
-                    &owner_class,
-                    method,
-                    &method_def,
-                    &cc,
-                    attributes,
-                    args,
-                    invocant,
-                    &empty_fns,
-                );
-                if pushed_dispatch {
-                    self.pop_method_dispatch();
-                }
-                self.pop_method_samewith_context();
-                let (result, new_attrs, attrs_adjusted) = method_result?;
-                if let Some(id) = target_id {
-                    // Commit only a `:=`-adjusted snapshot (cell-CAS race
-                    // avoidance — see the primary dispatch site).
-                    if attrs_adjusted && let Some(cell) = &attrs_cell {
-                        cell.commit_attrs(new_attrs.clone());
-                    }
-                    if !self.in_lvalue_assignment
-                        && let Value::Proxy { ref fetcher, .. } = result
-                    {
-                        // Without a `:=` adjustment the triple's map is the stale
-                        // entry snapshot (lazy reconcile) — re-snapshot the live
-                        // cell for the proxy fetcher.
-                        let proxy_attrs = if !attrs_adjusted && let Some(cell) = &attrs_cell {
-                            cell.to_map()
-                        } else {
-                            new_attrs
-                        };
-                        return loan_env!(self, proxy_fetch(fetcher, None, &cn, &proxy_attrs, id));
-                    }
-                }
-                return Ok(result);
-            }
-        }
-        // Native `.map` / `.grep` over a concrete array with a simple block (lever A).
-        // `target_name` is the receiver variable, enabling rw-binding writeback
-        // for `$_`-mutating blocks (`@a.map({ $_++ })` mutates `@a`).
-        if let Some(result) = self.try_native_array_map(Some(target_name), &target, method, &args) {
-            return result;
-        }
-        // Native `.subst` over a Str with a simple pattern/replacement (lever A).
-        if let Some(result) = self.try_native_subst(&target, method, &args) {
-            return result;
-        }
-        // Native `.sort` over a plain array with no/simple comparator (lever A).
-        if let Some(result) = self.try_native_sort(&target, method, &args) {
-            return result;
-        }
-        // Native `.min` / `.max` over a plain list, including `:by` blocks.
-        if let Some(result) = self.try_native_extrema(&target, method, &args) {
-            return result;
-        }
-        // Native `.minmax` over a plain list, including `:by` blocks.
-        if let Some(result) = self.try_native_minmax(&target, method, &args) {
-            return result;
-        }
-        // Native `.first` over a plain list (no-adverb forms), including blocks.
-        if let Some(result) = self.try_native_first(&target, method, &args) {
-            return result;
-        }
-        // Native QuantHash coercion `.Set`/`.Bag`/`.Mix`/`.SetHash`/`.BagHash`/
-        // `.MixHash` over a list-like aggregate or plain Cool scalar. Variable
-        // receivers (`@a.Set`, `$s.Bag`) compile to CallMethodMut and so land on this
-        // mut path; the
-        // coercion produces a *new* Set/Bag/Mix value and never mutates the
-        // receiver variable, so there is no writeback — identical to the non-mut
-        // path's native dispatch. Instance/Package receivers fall through.
-        if args.is_empty()
-            && let Some(result) = Self::try_native_quanthash_coerce(&target, method)
-        {
-            return result;
-        }
-        // Native `.Map` / `.Hash` coercion for variable receivers (`%h.Map`,
-        // `@a.Hash`) — same pure value op as the non-mut path, no writeback.
-        if args.is_empty()
-            && let Some(result) = Self::try_native_map_hash_coerce(&target, method)
-        {
-            return result;
-        }
-        // Native `.Seq` coercion for variable receivers (`@a.Seq`) — structural
-        // receivers only, same pure value op as the non-mut path.
-        if args.is_empty()
-            && method == "Seq"
-            && let Some(result) = crate::builtins::seq_coerce::to_seq_structural(&target)
-        {
-            return Ok(result);
-        }
-        // Native `.IO` coercion over a Cool scalar for variable receivers
-        // (`$s.IO`) — builds a *new* IO::Path and never mutates the receiver, so no
-        // writeback. Instance / non-IO Package / aggregate receivers fall through.
-        if let Some(result) = self.try_native_io_coercion(&target, method, &args) {
-            return result;
-        }
-        // Native `.encode` (Cool scalar -> Buf) / `.decode` (Buf/Blob -> Str) for
-        // variable receivers (`$s.encode("utf-16")`) — same pure transformation as
-        // the non-mut path; returns a *new* Buf/Str, no writeback.
-        if let Some(result) = self.try_native_encode_decode(&target, method, &args) {
-            return result;
-        }
-        // TODO: compile to bytecode — native/Buf/Failure method fork, mut (ledger §1).
-        // User-defined methods run as bytecode (compiled at registration or on
-        // demand above); what remains is native receiver dispatch blocked on
-        // ③ state ownership / first-class container Phase 2.
-        crate::vm::vm_stats::record_method_fallback(method);
-        self.vm_call_method_mut_with_values(target_name, target, method, args)
-    }
-
     /// Execute a protect block inline in the current Interpreter, avoiding the overhead
     /// of creating a new Interpreter.
     pub(super) fn exec_protect_block_inline(
@@ -2592,7 +226,7 @@ impl Interpreter {
 
     /// Check if a method candidate has a wrap chain from ^lookup().candidates[N].wrap().
     /// If so, dispatch through the wrapper and return Some(result).
-    fn check_method_wrap_chain(
+    pub(crate) fn check_method_wrap_chain(
         &mut self,
         cn: &str,
         owner_class: &str,
